+++ src/redturtle/prenotazioni/content/prenotazioni_folder.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
+from typing import Generator
+
 from collective.z3cform.datagridfield.datagridfield import DataGridFieldFactory
 from collective.z3cform.datagridfield.row import DictRow
 from plone.app.textfield import RichText
@@ -11,13 +13,14 @@ from z3c.form.browser.checkbox import CheckBoxFieldWidget
 from zope import schema
 from zope.component import provideAdapter
 from zope.i18n import translate
-from zope.interface import Interface, Invalid, implementer, invariant, provider
+from zope.interface import Invalid, implementer, invariant, provider
 from zope.schema.interfaces import IContextAwareDefaultFactory
 from zope.schema.vocabulary import SimpleTerm, SimpleVocabulary
 
 from redturtle.prenotazioni import _
 from redturtle.prenotazioni.browser.widget import WeekTableOverridesFieldWidget
 from redturtle.prenotazioni.config import DEFAULT_VISIBLE_BOOKING_FIELDS
+from redturtle.prenotazioni.content.booking_type import BookingType
 from redturtle.prenotazioni.content.validators import PauseValidator, checkOverrides
 
 try:
@@ -120,20 +123,6 @@ class IPauseTableRow(model.Schema):
         title=_("pause_end_label", default="Pause end"),
         vocabulary="redturtle.prenotazioni.VocOreInizio",
         required=False,
-    )
-
-
-class IBookingTypeRow(Interface):
-    name = schema.TextLine(title=_("Booking type name"), required=True)
-    duration = schema.Choice(
-        title=_("Duration value"),
-        required=True,
-        vocabulary="redturtle.prenotazioni.VocDurataIncontro",
-    )
-    hidden = schema.Bool(
-        title=_("Hidden type"),
-        required=False,
-        default=False,
     )
 
 
@@ -441,25 +430,6 @@ class IPrenotazioniFolder(model.Schema):
         ),
     )
 
-    booking_types = schema.List(
-        title=_("booking_types_label", default="Booking types"),
-        description=_(
-            "booking_types_help",
-            default="Put booking types there (one per line).\n"
-            "If you do not provide this field, "
-            "not type selection will be available. "
-            "If the 'Hidden Type' flag is selected the type will only "
-            "be available to users with the 'Bookings Manager' permission",
-        ),
-        value_type=DictRow(schema=IBookingTypeRow),
-    )
-    form.widget(
-        "booking_types",
-        DataGridFieldFactory,
-        auto_append=False,
-        frontendOptions={"widget": "data_grid"},
-    )
-
     gates = schema.List(
         title=_("gates_label", "Gates"),
         description=_("gates_help", default="Put gates here (one per line)."),
@@ -494,8 +464,6 @@ class IPrenotazioniFolder(model.Schema):
         """
         Needed because is the only way to validate a datagrid field
         """
-        if not data.booking_types:
-            raise Invalid(_("You should set at least one booking type."))
         for interval in data.week_table:
             if interval["morning_start"] and not interval["morning_end"]:
                 raise Invalid(_("You should set an end time for morning."))
@@ -765,3 +733,6 @@ class PrenotazioniFolder(Container):
 
     def getCosaServe(self):
         return self.cosa_serve
+
+    def get_booking_types(self) -> Generator[BookingType, None, None]:
+        return self.listFolderContents(contentFilter={"portal_type": "BookingType"})