+++ src/redturtle/prenotazioni/restapi/serializers/adapters/booking_type.py
@@ -1,0 +1,25 @@
+# -*- coding: utf-8 -*-
+from zope.component import adapter
+from zope.interface import implementer
+
+from redturtle.prenotazioni.content.booking_type import IBookingType
+from redturtle.prenotazioni.interfaces import (
+    IRedturtlePrenotazioniLayer,
+    ISerializeToRetroCompatibleJson,
+)
+
+
+@implementer(ISerializeToRetroCompatibleJson)
+@adapter(IBookingType, IRedturtlePrenotazioniLayer)
+class BookingTypeRetroCompatibleSerializer:
+    """BookingType c.t. serializer retrocompatible with earlier frontend implementation"""
+
+    def __init__(self, context, request):
+        self.context = context
+        self.request = request
+
+    def __call__(self, *args, **kwargs):
+        return {
+            "name": self.context.title,
+            "duration": str(self.context.duration),
+        }