+++ src/redturtle/prenotazioni/tests/test_prenotazione_email_sent_to_managers.py
@@ -42,11 +42,17 @@ class TestEmailToManagers(unittest.TestCase):
             title="Prenota foo",
             description="",
             daData=date.today(),
-            booking_types=[
-                {"name": "Type A", "duration": "30"},
-            ],
             gates=["Gate A"],
         )
+
+        api.content.create(
+            type="BookingType",
+            title="Type A",
+            duration=30,
+            container=self.folder_prenotazioni,
+            gates=["all"],
+        )
+
         week_table = self.folder_prenotazioni.week_table
         for data in week_table:
             data["morning_start"] = "0700"